requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

# 모델 리포의 fast 토크나이저를 쓰기 위한 호환 조합
transformers==4.44.2
tokenizers==0.20.1
huggingface-hub==0.24.6

python-dotenv==1.0.1
pydantic==2.7.3

# CPU PyTorch
torch==2.5.1
